Tidying up intro/methods sections
Lots of results section changes. Main focus was switching to reporting urate and related genes that had a haplotypic result. List of focus criteria at top of genes section

Aiming to really tighten up the results to be pretty much only focussed on the specific genes

Thinking about doing a substantial change to the first part of the results, moving the stat info to supplementary and generally moving the specific stat subsections to be relevant paragraphs under the 'Selection in Polynesians' section